move util func to Bug class
